UpdateNotes
Atualização do container form

+++ Códigos/script.js
@@ -1,22 +1,22 @@
-// Fecchar Popup
-
 const popup = document.querySelector('.popup'),
 formTitulo = popup.querySelector('.header-popup h3'),
 iconFechar = popup.querySelector('i'),
 inputTitulo = document.getElementById('title-input'),
 inputDescricao = document.getElementById('note-input'),
 buttonAdd = document.getElementById('button-add'),
-addNote = document.getElementById('add'),
+addBtn = document.getElementById('add-box'),
 notes = JSON.parse(localStorage.getItem("Notação:") || "[]");
+
+const addCheck = document.getElementById('add-check');
 
 let update  = false, updtadid;
 
 iconFechar.addEventListener('click', fecharPopup);
-addNote.addEventListener('click', abrirPopup);
+addBtn.addEventListener('click', abrirPopup)
 buttonAdd.addEventListener('click', enviarNote);
 
 // Fechar janela
-function fecharPopup()
+function fecharPopup()  
 {
 	popup.classList.remove('open')
 	inputTitulo.value = "";
@@ -39,14 +39,26 @@ function tagNote()
 	notes.forEach((note, id) =>
 	{
 		let boxNote = 
-		'<div class="note"> <div class="header-note"><h3>'+note.title+'</h3>'+
-		'<p>'+note.descricao+'</p>'+
-		'</div><div class="footer-note">'+
-		'<div class="date">Day, hora</div><div class="edite-note"><i class="bx bx-edit-alt"></i><div class="option-edite">'+
-		'<button onclick="atualizar('+id+')">Editar</button>'+
-		'<button onclick="deletar('+id+')">Excluir</button></div></div></div></div>';
+		`
+		<div class="note"> 
+			<div class="header-note">
+				<h3>${note.title}</h3>
+				<p>${note.descricao}<p>
+			</div>
+			<div class="footer-note">
+				<div class="date">Data: ${note.data}</div>
+				<div class="edite-note">
+					<i class="bx bx-edit-alt"></i>
+					<div class="option-edite">
+						<button onclick="atualizar(${id}, '${note.title}', '${note.descricao}')">Editar</button>
+						<button onclick="deletar(${id})">Excluir</button>
+					</div>
+				</div>
+			</div>
+		</div>
+		`;
 
-		addNote.insertAdjacentHTML("afterend", boxNote);
+		addBtn.insertAdjacentHTML("afterend", boxNote);
 	})
 };
 tagNote();
@@ -57,10 +69,18 @@ function enviarNote()
 
 	let tituloNote = inputTitulo.value;
 	let descricaoNote = inputDescricao.value;
+	let data = new Date();
+	let mes = data.getMonth();
+	let dia = data.getDate();
+
+	mes = mes < 10 ? "0" + mes : mes;
+	dia = dia < 10 ? "0" + dia : dia;
+
+	let dataNote = dia +" / "+ mes;
 
 	if (tituloNote || descricaoNote) 
 	{
-		let infor = { title: tituloNote, descricao: descricaoNote};
+		let infor = { title: tituloNote, descricao: descricaoNote, data: dataNote};
 
 		if (!update) 
 		{
@@ -81,19 +101,28 @@ function enviarNote()
 // Deletar notação
 function deletar(id)
 {
-	notes.splice(id, 1);
+	let aviso = confirm("Tem certeza?")
 
-	tagNote()
-	localStorage.setItem("Notação:", JSON.stringify(notes))
+	if(!aviso) return
+	{
+		notes.splice(id, 1);
+
+		tagNote();
+		localStorage.setItem("Notação:", JSON.stringify(notes));
+	}
+
 }
 
-function atualizar(id)
+function atualizar(id, titulo, desc)
 {
 	update  = true;
 	updtadid = id;
 
+	inputTitulo.value = titulo;
+	inputDescricao.value = desc;
+
 	formTitulo.innerText = "Atualizar Notação";
 
-	addNote.click();
+	addBtn.click();
 	localStorage.setItem("Notação:", JSON.stringify(notes));	
 }
